requirements.txt
gunicorn
Flask
tensorflow==2.3.0
numpy
opencv-python-headless
pillow